Normalize arrangement
