+++ lib/normalize/harmony.js
