+++ lib/normalize/v2.js
