+++ lib/parse.js
@@ -5,15 +5,6 @@ const assert = require('assert');
 const nodeEval = require('node-eval');
 
 const detect = require('./detect');
-const normalize = require('./normalize');
-const normalize2 = require('./normalize2');
-const normalizeHarmony = require('./normalize-harmony');
-
-const normalizers = {
-    v1: normalize,
-    v2: normalize2,
-    harmony: normalizeHarmony
-};
 
 /**
  * Parses BEMDECL file data
@@ -24,22 +15,31 @@ const normalizers = {
 module.exports = function parse(bemdecl) {
     assert(typeof bemdecl === 'object' || typeof bemdecl === 'string', 'Bemdecl must be String or Object');
 
+    const normalize = require('.').normalize;
+
     const declaration = (typeof bemdecl === 'string') ? nodeEval(bemdecl) : bemdecl;
     const hasOwn = Object.prototype.hasOwnProperty.bind(Object(declaration));
     const format = declaration.format || detect(declaration);
 
+    let decl;
+
     switch (format) {
         case 'v1':
             assert(hasOwn('blocks'), 'Invalid declaration format');
-            return normalizers.v1(declaration.blocks);
+            decl = declaration.blocks;
+            break;
         case 'v2':
         case 'enb':
             assert(hasOwn('decl') || hasOwn('deps'), 'Invalid format of declaration.');
-            return normalizers.v2(declaration.decl || declaration.deps);
+            decl = declaration.decl || declaration.deps;
+            break;
         case 'harmony':
             assert(hasOwn('decl'), 'Invalid format of declaration.');
-            return normalizers.harmony(declaration.decl);
+            decl = declaration.decl;
+            break;
         default:
             throw new Error('Unknown BEMDECL format.');
     }
+
+    return normalize(decl, { format: format });
 };
