+++ test/normalize2/elems-mod.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const test = require('ava');
-const normalize = require('../../lib/normalize2');
+const normalize = require('../../lib/normalize/v2');
 
 test('should support shortcut for bool mod of elem', t => {
     const decl = { block: 'block', elems: 'elem', mod: 'mod' };
